requirements.txt
# QuantumTrader Production - Requirements
# Python 3.12 compatible

# Core Data Science
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.0.0

# Trading & Finance
python-dotenv==1.0.0

# Visualization (for backtest)
matplotlib==3.8.0
seaborn==0.13.0

# Performance & Optimization (Optional)
numba>=0.61.0          # JIT compilation (Python 3.12+ requires 0.61+)
bottleneck>=1.5.0      # Fast numpy operations

# Monitoring & Logging
colorama==0.4.6        # For enhanced console monitor
psutil==5.9.5          # System monitoring

# Communication
pyzmq>=27.0.0          # For HMARL broadcasting (Python 3.12+ requires 27+)
msgpack>=1.1.0         # Message serialization
lz4>=4.4.0             # Compression

# Development Tools (optional)
pytest==7.4.0          # Testing
pytest-cov==4.1.0      # Coverage
black==23.7.0          # Code formatting
pylint==2.17.5         # Code linting

# Windows specific (if needed)
# pywin32==306         # Windows API access